Uses components from query :)
The parser now uses components from the query and the old component lock have been removed.
Functions from update_component.rs has been rewritten and transfered into send_query.rs.
