+++ src/ProtobufServer/ecdar_requests/mod.rs
@@ -1,2 +1,1 @@
 mod send_query;
-mod update_components;
