+++ src/ProtobufServer/ecdar_requests/send_query.rs
@@ -1,5 +1,10 @@
+use std::collections::HashMap;
 use std::panic::AssertUnwindSafe;
 
+use crate::component::Component;
+use crate::xml_parser::parse_xml_from_str;
+use crate::DataReader::component_loader::ComponentContainer;
+use crate::DataReader::json_reader::json_to_component;
 use crate::DataReader::json_writer::component_to_json;
 use crate::DataReader::parse_queries;
 use crate::ModelObjects::queries::Query;
@@ -10,9 +15,11 @@ use crate::ProtobufServer::services::query_response::query_ok::{
 };
 use crate::ProtobufServer::services::query_response::QueryOk;
 use crate::ProtobufServer::services::query_response::Response as QueryOkOrErrorResponse;
-use crate::ProtobufServer::services::{Component, QueryRequest, QueryResponse};
+use crate::ProtobufServer::services::{
+    Component as ProtobufComponent, QueryRequest, QueryResponse,
+};
 use crate::System::executable_query::QueryResult;
-use crate::System::extract_system_rep;
+use crate::System::{extract_system_rep, input_enabler};
 use log::trace;
 use tonic::{Request, Response, Status};
 
@@ -25,11 +32,20 @@ impl ConcreteEcdarBackend {
     ) -> Result<Response<QueryResponse>, Status> {
         trace!("Received query: {:?}", request);
         let query_request = request.0.into_inner();
-
+        let components_info = query_request.components_info.as_ref().unwrap();
+        let proto_components = &components_info.components;
         let query = parse_query(&query_request)?;
 
-        let components = self.get_components_lock()?;
-        let mut component_container = components.borrow_mut();
+        let mut parsed_components = vec![];
+
+        for proto_component in proto_components {
+            let components = parse_components_if_some(proto_component)?;
+            for component in components {
+                parsed_components.push(component);
+            }
+        }
+
+        let mut component_container = create_component_container(parsed_components);
 
         if query_request.ignored_input_outputs.is_some() {
             return Err(Status::unimplemented(
@@ -38,7 +54,7 @@ impl ConcreteEcdarBackend {
         }
 
         let executable_query =
-            match extract_system_rep::create_executable_query(&query, &mut *component_container) {
+            match extract_system_rep::create_executable_query(&query, &mut component_container) {
                 Ok(query) => query,
                 Err(e) => {
                     return Err(Status::invalid_argument(format!(
@@ -72,6 +88,50 @@ fn parse_query(query_request: &QueryRequest) -> Result<Query, Status> {
     }
 }
 
+fn parse_components_if_some(
+    proto_component: &ProtobufComponent,
+) -> Result<Vec<Component>, tonic::Status> {
+    if let Some(rep) = &proto_component.rep {
+        match rep {
+            Rep::Json(json) => parse_json_component(json),
+            Rep::Xml(xml) => Ok(parse_xml_components(xml)),
+        }
+    } else {
+        Ok(vec![])
+    }
+}
+
+fn parse_json_component(json: &str) -> Result<Vec<Component>, tonic::Status> {
+    match json_to_component(json) {
+        Ok(comp) => Ok(vec![comp]),
+        Err(_) => Err(tonic::Status::invalid_argument(
+            "Failed to parse json component",
+        )),
+    }
+}
+
+fn parse_xml_components(xml: &str) -> Vec<Component> {
+    let (comps, _, _) = parse_xml_from_str(xml);
+    comps
+}
+
+fn create_component_container(components: Vec<Component>) -> ComponentContainer {
+    let mut comp_hashmap = HashMap::<String, Component>::new();
+    for mut component in components {
+        trace!("Adding comp {} to container", component.get_name());
+
+        component.create_edge_io_split();
+        let inputs: Vec<_> = component
+            .get_input_actions()
+            .into_iter()
+            .map(|channel| channel.name)
+            .collect();
+        input_enabler::make_input_enabled(&mut component, &inputs);
+        comp_hashmap.insert(component.get_name().to_string(), component);
+    }
+    ComponentContainer::new(comp_hashmap)
+}
+
 fn convert_ecdar_result(query_result: &QueryResult) -> Option<ProtobufResult> {
     match query_result {
         QueryResult::Refinement(refines) => Some(ProtobufResult::Refinement(RefinementResult {
@@ -80,7 +140,7 @@ fn convert_ecdar_result(query_result: &QueryResult) -> Option<ProtobufResult> {
             state: None, // TODO: this is a placeholder value
         })),
         QueryResult::GetComponent(comp) => Some(ProtobufResult::Component(ComponentResult {
-            component: Some(Component {
+            component: Some(ProtobufComponent {
                 rep: Some(Rep::Json(component_to_json(comp))),
             }),
         })),
