+++ src/tests/grpc/mod.rs
@@ -1,2 +1,1 @@
 pub mod send_query;
-pub mod update_components;
